netbsd compile fix
svn path=/trunk/; revision=290

+++ pan/data/parts.cc
@@ -80,7 +80,7 @@ namespace
     register const char *k, *m;
     const StringView& key (key_mid.to_view());
 
-    const int bmax = std::min ((int)std::min (key.len, mid.len), UCHAR_MAX);
+    const int bmax = std::min ((int)std::min (key.len, mid.len), (int)UCHAR_MAX);
     k = &key.front();
     m = &mid.front();
     for (; b!=bmax; ++b)
